Show anno in anno windows when "show" is clicked
From the hover menu in the image window, when the user clicks the newly-created "show" link, the annotation windows will scroll to and highlight the corresponding annotation. If no existing annotation windows contain the annotation, a new window will be created.

This commit also contains some conversions from jQuery Deferred to ES6 Promise.